requirements.txt
alembic==1.0.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pkg-resources==0.0.0
pluggy==0.7.1
py==1.6.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.7.4
pytest-flask==0.11.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.11
typed-ast==1.1.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
